Move utilities that will be shared into their own module
