+++ test/lint/test-packages.js
@@ -3,8 +3,14 @@
 
 import expect from 'expect';
 import fs from 'fs-promise';
-import glob from 'glob';
 import difference from 'lodash/difference';
+import {
+  REQUIRES_REGEX,
+  IMPORTS_REGEX,
+  autoFailingAsyncTest,
+  globMany,
+  regexFiles,
+} from './shared.js';
 
 const all = '/**/';
 const valid = '*.@(js|jsx)';
@@ -36,13 +42,8 @@ const ignored = [
   'eslint-plugin-react',
 ];
 
-// We can afford being a little more relaxed with the regexes here
-// since we have very strict eslint rules about whitespace.
-const requires = /require\('([\w-/]+)'\)/g;
-const imports = /import(?:.*?from)?\s+'([\w-/]+)'/g;
-
 describe('packages', () => {
-  it('should not be orpahned', test(async function() {
+  it('should not be orphaned', autoFailingAsyncTest(async function() {
     const manifest = await fs.readJson('package.json');
     const modules = [
       ...Object.keys(manifest.dependencies),
@@ -50,71 +51,9 @@ describe('packages', () => {
     ];
 
     const sources = await globMany(paths);
-    const matches = await regexFiles(sources, requires, imports);
+    const matches = await regexFiles(sources, REQUIRES_REGEX, IMPORTS_REGEX);
     const expected = Array.from(new Set([...matches, ...ignored]));
 
     expect(difference(modules, expected)).toEqual([]);
   }));
 });
-
-function test(runner) {
-  return async function(done) {
-    let caught;
-    try {
-      await runner();
-    } catch (err) {
-      caught = err;
-    }
-    done(caught);
-  };
-}
-
-function globOne(wildcard) {
-  return new Promise((resolve, reject) => {
-    glob(wildcard, {}, (err, files) => {
-      if (err) {
-        reject(err);
-        return;
-      }
-      resolve(files);
-    });
-  });
-}
-
-async function globMany(wildcards) {
-  let out = [];
-  for (const wildcard of wildcards) {
-    const files = await globOne(wildcard);
-    out = out.concat(files);
-  }
-  return out;
-}
-
-async function regexSingleFile(path, regexps) {
-  const contents = await fs.readFile(path, 'utf8');
-  let out = [];
-
-  for (const regex of regexps) {
-    let match = regex.exec(contents);
-
-    while (match) {
-      const [, ...captures] = match;
-
-      // In the case of submodule imports e.g. 'lodash/throttle', we're only
-      // interested in the main module path component.
-      const modules = captures.map(e => e.split('/')[0]);
-      out = out.concat(modules);
-      match = regex.exec(contents);
-    }
-  }
-  return out;
-}
-
-async function regexFiles(files, ...regexps) {
-  let out = [];
-  for (const file of files) {
-    const matches = await regexSingleFile(file, regexps);
-    out = out.concat(matches);
-  }
-  return Array.from(new Set(out));
-}
